prevent field browser calls when it's not loaded
